refactor(storage): extract the bulk-read surfaces into storage/memory-read-store.ts
God-file decomposition, second storage cluster. Moves the bulk/windowed
read surfaces verbatim (662 LOC, 10 methods): collectActiveMemoryPaths,
readMemoriesWindow, readAllColdMemories, readMemoryByPath,
readAllArtifactsCached, readQuestions, buffer-surprise ledger
read/append, compression-guideline state reads, wearable transcript-day
listing — into MemoryReadStore (26 deps incl. 8 dir getters, all live
via selfDeps). Cache fields/statics the store reads widened from
private with pointer comments.

storage.ts: 7,227 -> 6,650 LOC (7,747 at program start).
